move in-process startup to a strategy class
